Style: Change RabbitMQ Setting File
